chore: trim list of west modules
Bridle imports almost all modules from upstream zephyr resulting in long setup times when for the purpose of the dojo only a very small number of modules is actually needed.

Instead of relying on the bridle manifest with this commit zephyr is brought in directly with a much smaller allow-list.